为SSTManage添加mx_sequenceNumber成员，为了恢复没来得及写入SSTable的数据。并修改相关函数。 修改了SST的相关函数writeSST(); 开放了 SSTManage的枚举变量为public。

+++ src/SSTable.h
@@ -179,8 +179,8 @@ public:
     ~SSTable(){if(_fileDescriptor>=0){close(_fileDescriptor);}}
     void writeBlockIndex();
     void writeFooter();
-    int writeSST(SkipList::writeIterator iterator);
-    int writeSST(vector<Data>::iterator begin,vector<Data>::iterator end);
+    ULL writeSST(SkipList::writeIterator iterator);
+    ULL writeSST(vector<Data>::iterator begin,vector<Data>::iterator end);
     void setMinMax(){
         assert(_indexVector.size());
         _miniData=_indexVector[0].minData;
